Fix error in ButtonCreator script
The ButtonCreator script was trying to assign a List<string> to a Dictionary<string, System.Action>, which caused a compilation error. This commit changes the type of the variable and modifies the GetAvailablePrompts method to use the GetAvailablePrompts method of the ZimObject class instead of the actions dictionary. This commit also removes the actions dictionary from the ZimObject subclass, as it is no longer needed.